+++ src/CSET/cset_workflow/rose-suite.conf.example
@@ -22,7 +22,6 @@ COLORBAR_FILE=""
 !!CONDA_METPLUS_VENV_LOCATION=""
 CONDA_PATH=""
 CONTROL_MEMBER=0
-COORDINATE_LIST=""
 CSET_CASE_DATES=[]
 CSET_CYCLING_MODE="case_study"
 CSET_ENV_SEPARATE_MET=False
