+++ tests/conftest.py
@@ -17,14 +17,17 @@
 https://docs.pytest.org/en/latest/reference/fixtures.html#conftest-py-sharing-fixtures-across-multiple-files
 """
 
+import datetime
 import shutil
 import subprocess
 from collections.abc import Callable
 from pathlib import Path
 from uuid import uuid4
 
+import cf_units
 import iris
 import iris.cube
+import numpy as np
 import pytest
 
 from CSET.operators import constraints, filters, read
@@ -1031,3 +1034,54 @@ def precalc_wind_below_maul_5d_read_only():
 def precalc_wind_below_maul_5d(precalc_wind_below_maul_5d_read_only):
     """Get precalculated wind below maul for 5D data. It is safe to modify."""
     return precalc_wind_below_maul_5d_read_only.copy()
+
+
+@pytest.fixture
+def feature_cube() -> iris.cube.Cube:
+    """Set up three timesteps and three realizations of data and place into cube."""
+    data_arr = np.zeros((3, 3, 10, 10))
+    data_arr[0:2, 0, 2:6, 2:6] = 1
+    data_arr[0:2, 1, 3:7, 3:7] = 1
+    data_arr[0:2, 2, 4:8, 4:8] = 1
+
+    realization = iris.coords.DimCoord(points=[0, 1, 2], standard_name="realization")
+    time_units = cf_units.Unit("days since 2000-01-01 00:00:00", calendar="gregorian")
+    time_start = datetime.datetime(2010, 1, 1, 0, 0, 0)
+    time_dt_points = [
+        time_start + datetime.timedelta(minutes=5 * idx) for idx in range(3)
+    ]
+    time_points = [time_units.date2num(time_point) for time_point in time_dt_points]
+    time_coord = iris.coords.DimCoord(
+        points=time_points, standard_name="time", units=time_units
+    )
+
+    coord_system = iris.coord_systems.TransverseMercator(
+        latitude_of_projection_origin=55, longitude_of_central_meridian=0
+    )
+    coord_range = np.arange(0, 100, 10)
+    proj_y_coord = iris.coords.DimCoord(
+        points=coord_range,
+        standard_name="projection_y_coordinate",
+        var_name="projection_y_coordinate",
+        units="m",
+        coord_system=coord_system,
+    )
+    proj_x_coord = iris.coords.DimCoord(
+        points=coord_range,
+        standard_name="projection_x_coordinate",
+        var_name="projection_x_coordinate",
+        units="m",
+        coord_system=coord_system,
+    )
+
+    proj_y_coord.guess_bounds()
+    proj_x_coord.guess_bounds()
+
+    coords = (realization, time_coord, proj_y_coord, proj_x_coord)
+    dim_coords_and_dims = [(coord, dim) for dim, coord in enumerate(coords)]
+    cube = iris.cube.Cube(
+        data=data_arr,
+        dim_coords_and_dims=dim_coords_and_dims,
+        long_name="crps test",
+    )
+    return cube
